refactor(checker): split async Promise diagnostics out of function_type_helpers.rs
- extract TS2468/TS2705/TS1055/TS1064 async-Promise checks into a child module (2019 -> 1583 lines) to satisfy the 2000-line arch guard
